Refactor Convey to FeedShooter
